+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## UNRELEASED
 
+### Changes
+
+* Intercept 403 Vite middleware responses to provide helpful error messages for host validation issues.
+
+### Fixes
+
 * Fixes for native Windows Node.js without WSL.
 
 ## 1.1.0 (2025-06-11)